Space Station Updated
Added a Space station. Spawns on 20% of tiles, and always on the first tile within 1000 pixels of spawn.  Going near brings up a prompt to "press e", and pressing "e" will bring you to the shop. Prompt just prints "press e" and shop does not yet exits, but the buttons to get to them work.