Bitmojis, Full Store Functionally, HUD
Adjusted Aishas bitmoji code taken from her github webpage to allow for use in the project. Added full store functionality with three upgrades. Added a Hud that tracks player coordinates on a tile, which tile they are on, and if there is a space station that is indicated and its location is shown. 

Changed settings menu to an Instructions page.  
Changed which scene is which for organisational reasons.